requirements.txt
Django
django-dotenv
django-filter
djangorestframework
Markdown
psycopg2
python-dateutil
pytz
six
StringGenerator
uWSGI
wheel
